scene: refine rendering and window integration

+++ scene/wlf_ellipse_node.c
@@ -21,7 +21,7 @@ static void bounds(const struct wlf_ellipse_shape *shape,
 }
 
 struct wlf_ellipse_node *wlf_ellipse_node_create(struct wlf_scene_node *parent,
-		double x, double y, struct wlf_ellipse_shape *shape) {
+		int x, int y, struct wlf_ellipse_shape *shape) {
 	if (shape == NULL) return NULL;
 	struct wlf_ellipse_node *node = calloc(1, sizeof(*node));
 	if (node == NULL) return NULL;
@@ -57,7 +57,7 @@ static void render_at(struct wlf_ellipse_node *node,
 void wlf_ellipse_node_render(struct wlf_ellipse_node *node,
 		struct wlf_ellipse_pass *pass, struct wlf_render_target_info *target,
 		const pixman_region32_t *clip) {
-	double x, y;
+	int x, y;
 	if (node == NULL || !wlf_scene_node_coords(&node->base, &x, &y)) {
 		return;
 	}
