+++ scene/wlf_shape_node_common.c
@@ -4,7 +4,7 @@
 
 bool wlf_shape_node_common_init(struct wlf_scene_node *node,
 		const struct wlf_scene_node_impl *impl, struct wlf_scene_node *parent,
-		double x, double y, double minx, double miny, double maxx, double maxy) {
+		int x, int y, double minx, double miny, double maxx, double maxy) {
 	if (node == NULL || impl == NULL || parent == NULL ||
 			!isfinite(minx) || !isfinite(miny) || maxx < minx || maxy < miny) {
 		return false;
@@ -12,8 +12,8 @@ bool wlf_shape_node_common_init(struct wlf_scene_node *node,
 	wlf_scene_node_init(node, impl, parent);
 	node->state.x = x;
 	node->state.y = y;
-	node->state.width = ceil(maxx - minx);
-	node->state.height = ceil(maxy - miny);
+	node->state.width = (uint32_t)ceil(maxx - minx);
+	node->state.height = (uint32_t)ceil(maxy - miny);
 	return true;
 }
 
@@ -23,15 +23,15 @@ bool wlf_shape_node_common_refresh_at(struct wlf_scene_node *node,
 	if (!isfinite(minx) || !isfinite(miny) || maxx < minx || maxy < miny) {
 		return false;
 	}
-	node->state.width = ceil(maxx - minx);
-	node->state.height = ceil(maxy - miny);
+	node->state.width = (uint32_t)ceil(maxx - minx);
+	node->state.height = (uint32_t)ceil(maxy - miny);
 	*offset_x = x - minx;
 	*offset_y = y - miny;
 	return true;
 }
 
 void wlf_shape_node_common_get_size(struct wlf_scene_node *node,
-		double *width, double *height) {
+		uint32_t *width, uint32_t *height) {
 	*width = node->state.width;
 	*height = node->state.height;
 }
@@ -59,16 +59,16 @@ struct wlf_scene_node *wlf_shape_node_common_at(struct wlf_scene_node *node,
 }
 
 void wlf_shape_node_common_bounds(struct wlf_scene_node *node,
-		double x, double y, pixman_region32_t *visible) {
+		int x, int y, pixman_region32_t *visible) {
 	if (!wlf_shape_node_common_invisible(node))
-		pixman_region32_union_rect(visible, visible, (int)x, (int)y,
-			(uint32_t)node->state.width, (uint32_t)node->state.height);
+		pixman_region32_union_rect(visible, visible, x, y,
+			node->state.width, node->state.height);
 }
 
 bool wlf_shape_node_common_in_box(struct wlf_scene_node *node,
 		struct wlf_frect *box, scene_node_box_iterator_func_t iterator, void *data) {
 	if (wlf_shape_node_common_invisible(node)) return false;
-	double x, y;
+	int x, y;
 	if (!wlf_scene_node_coords(node, &x, &y)) return false;
 	if (x >= box->x + box->width || x + node->state.width <= box->x ||
 			y >= box->y + box->height || y + node->state.height <= box->y) return false;
@@ -76,6 +76,6 @@ bool wlf_shape_node_common_in_box(struct wlf_scene_node *node,
 }
 
 bool wlf_shape_node_common_construct_render_list_iterator(
-		struct wlf_scene_node *node, double lx, double ly, void *data) {
+		struct wlf_scene_node *node, int lx, int ly, void *data) {
 	return wlf_scene_node_add_render_list_entry(node, lx, ly, data);
 }
